db · Frente C · troceo 5/N · trocear initDB → upgrade-a/b + post-open + ensure-index · db.ts 1909→540
Última pieza (fiscal-crítica · protegida por el test network-first). Mover-no-reescribir:
- El callback `upgrade` (1080 líneas) se parte en dos por un límite de sentencia limpio
  (line 1008, `if(oldVersion<55)`; parte A sin ningún `return` → corre entera antes de B):
  services/db/upgrade-a.ts (createStores + migraciones tempranas) y upgrade-b.ts
  (migraciones oldVersion<55…75). db.ts hace `applyUpgradeA(...); return applyUpgradeB(...)`
  → se preserva la semántica de `return (async()=>…)()` (idb espera la promesa dentro de
  la versionchange tx). Copia literal de los cuerpos.
- Las 6 migraciones post-open (`dbPromise = dbPromise.then(...)`) → services/db/post-open.ts
  (runPostOpenMigrations · cuerpos idénticos). db.ts: `dbPromise = runPostOpenMigrations(dbPromise)`.
- El helper `ensureIndex` → services/db/ensure-index.ts (lo usan upgrade-a/b).

db.ts 1909→540. Todos los ficheros de services/db/ <800 (máx types-inmuebles 634). tsc 0.
Verificación fiscal-crítica: network-first (45 stores+índices en apertura limpia a través
del initDB troceado) · initDBBloque3 (limpio nuevo + v79 existente) · dbV78Alquileres
(migración post-open) VERDES. dbV60/dbV65 rojos son PRE-EXISTENTES (verificado con stash
en el commit pre-troceo · idénticos 6 fallos).
